# backend_school_crm/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.SubscriptionPayment:
    properties:
      amount:
        type: number
      click_payment_id:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      invoice_number:
        type: string
      notes:
        type: string
      payment_date:
        type: string
      payment_method:
        type: string
      status:
        description: pending, completed, failed, refunded
        type: string
      stripe_payment_id:
        type: string
      subscription_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@school-crm.example
    name: School CRM Support
  description: Residual monolith — auth, subscriptions, Click.uz/Telegram payments,
    messaging, reports, dev panel.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: School CRM API
  version: "1.0"
paths:
...
  /payments:
    get:
      description: Returns payments for a branch, with optional month/year/status
        filters.
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        required: true
        type: string
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year (e.g. 2025)
        in: query
        name: year
        type: integer
      - description: paid | partial | not_paid
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List payments
      tags:
      - payments
    post:
      consumes:
      - application/json
      description: Records a new payment entry for a student.
      parameters:
      - description: student_id, branch_id, amount, month, year, payment_method
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: created payment
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create payment
      tags:
      - payments
  /payments/{id}:
    delete:
      description: Permanently removes a payment record.
      parameters:
      - description: Payment UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: payment deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete payment
      tags:
      - payments
    put:
      consumes:
      - application/json
      parameters:
      - description: Payment UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update payment
      tags:
      - payments
  /payments/consolidated/data:
    get:
      description: Returns payment totals, method breakdown, and per-student summary
        for the dashboard.
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        required: true
        type: string
      - description: Month (1-12)
        in: query
        name: month
        required: true
        type: integer
      - description: Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Consolidated payment data
      tags:
      - payments
...
  /subscriptions/{id}/payments:
    get:
      description: Retrieve payment history for a subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SubscriptionPayment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get subscription payments
      tags:
      - subscriptions
...
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"